Improving the way sonar filtering is done.  Use a low pass float filter so that we don't have step-changes in sonar distance which can lead to spikes on the speed calculation.

+++ ArduCopter/Log.pde
@@ -157,10 +157,10 @@ select_logs(uint8_t argc, const Menu::arg *argv)
 
 struct PACKED log_Sonar {
     LOG_PACKET_HEADER;
-    uint16_t sonar_reading;
+    float sonar_reading;
     int16_t graffiti_controll;
 };
-static void Log_Write_Sonar(uint16_t tsonar, int16_t tpitch)
+static void Log_Write_Sonar(float tsonar, int16_t tpitch)
 {
 
     struct log_Sonar pkt = {
